pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "cctx"
version = "0.1.0"
description = "Agentic context harvester — assembles high-signal docs into CONTEXT.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "watchfiles>=0.21",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-snapshot",
    "respx>=0.21",
    "pytest-asyncio>=0.23",
]

[project.scripts]
cctx = "cctx.cli:cli"

[tool.setuptools.packages.find]
include = ["cctx*"]
